Added allow_new=false flag to bface_mask.
With the default (allow_new=true) the code creates all interior facets and the selects those
which fit to the mask, instead of considering only the existing boundary facets.
This is quite some overhead in 3D and unnecessary here (I think). I consider this a design
bug in ExtendableGrids which needs to be fixed in a breaking version.